requirements.txt
selenium==4.17.2
webdriver-manager==4.0.1
Pillow==10.4.0
PyYAML==6.0.3
pytest==8.4.2
ruff==0.1.13
mypy==1.8.0
